Change `ValidateTaskProperties` task to also validate the properties of artifact transform actions and parameter types.
The task treats all problems with these types as errors, which is the same as the behaviour runtime. This is another step towards reporting on problems with all plugin types consistently at runtime and plugin build time, and later in other places too. There are still some task type checks that are reported only by the validation task and some artifact transform checks that are reported only at runtime. These can be cleaned up later.